refactor: organizing navigation styles
