+++ src/navigation/StackScreen/styles.ts
@@ -1,0 +1,10 @@
+import { colorPalette } from "../../theme/colors";
+
+export const clearHeaderOptions = {
+  headerBackVisible: false,
+  headerTitle: "",
+  headerShadowVisible: false,
+  headerStyle: {
+    backgroundColor: colorPalette.backgroundWhite,
+  },
+};
