+++ src/navigation/Tabs/styles.ts
@@ -1,0 +1,14 @@
+import { colorPalette } from "../../theme/colors";
+
+export const tabStyleOptions = {
+    headerShown: false,
+    tabBarActiveTintColor: colorPalette.primary,
+    tabBarInactiveTintColor: colorPalette.mediumLight,
+  
+    tabBarStyle: {
+      backgroundColor: colorPalette.backgroundWhite,
+      elevation: 0,
+      shadowOpacity: 0,
+      borderTopWidth: 0,
+    },
+  };
